+++ .idea/workspace.xml
@@ -4,9 +4,9 @@
     <list default="true" id="99292fb8-fc1c-4a2d-a504-0b1ea6c77fb2" name="Changes" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/je_auto_control/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/je_auto_control/__init__.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/je_auto_control/utils/action_executer/action_execute.py" beforeDir="false" afterPath="$PROJECT_DIR$/je_auto_control/utils/action_executer/action_execute.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/je_auto_control/utils/je_auto_control_exception/exception_tag.py" beforeDir="false" afterPath="$PROJECT_DIR$/je_auto_control/utils/je_auto_control_exception/exception_tag.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/je_auto_control/utils/je_auto_control_exception/exceptions.py" beforeDir="false" afterPath="$PROJECT_DIR$/je_auto_control/utils/je_auto_control_exception/exceptions.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/je_auto_control/utils/action_executer/action_execute.py" beforeDir="false" afterPath="$PROJECT_DIR$/je_auto_control/utils/action_executer/action_executor.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/je_auto_control/utils/action_file/json_loader.py" beforeDir="false" afterPath="$PROJECT_DIR$/je_auto_control/utils/action_file/json_file.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/je_auto_control/wrapper/auto_control_record.py" beforeDir="false" afterPath="$PROJECT_DIR$/je_auto_control/wrapper/auto_control_record.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -64,8 +64,8 @@
       <recent name="D:\WorkSpaces\Program WorkSpace\Python\Project\Python_JEAutoControl\test\platform_independent\test\image" />
     </key>
   </component>
-  <component name="RunManager" selected="Python.action_execute">
-    <configuration name="action_execute" type="PythonConfigurationType" factoryName="Python" temporary="true" nameIsGenerated="true">
+  <component name="RunManager" selected="Python.action_executor">
+    <configuration name="action_executor" type="PythonConfigurationType" factoryName="Python" temporary="true" nameIsGenerated="true">
       <module name="Python_JEAutoControl" />
       <option name="INTERPRETER_OPTIONS" value="" />
       <option name="PARENT_ENVS" value="true" />
@@ -78,7 +78,7 @@
       <option name="ADD_CONTENT_ROOTS" value="true" />
       <option name="ADD_SOURCE_ROOTS" value="true" />
       <EXTENSION ID="PythonCoverageRunConfigurationExtension" runner="coverage.py" />
-      <option name="SCRIPT_NAME" value="$PROJECT_DIR$/je_auto_control/utils/action_executer/action_execute.py" />
+      <option name="SCRIPT_NAME" value="D:\WorkSpaces\Program WorkSpace\Python\Project\Python_JEAutoControl\je_auto_control\utils\action_executer\action_executor.py" />
       <option name="PARAMETERS" value="" />
       <option name="SHOW_COMMAND_LINE" value="false" />
       <option name="EMULATE_TERMINAL" value="false" />
@@ -221,15 +221,15 @@
     </configuration>
     <list>
       <item itemvalue="Python.test_auto_control_exception" />
-      <item itemvalue="Python.action_execute" />
       <item itemvalue="Python.critical_exit_test" />
       <item itemvalue="Python.execute_action_test" />
       <item itemvalue="Python.mouse_test" />
       <item itemvalue="Python.keyboard_write_test" />
+      <item itemvalue="Python.action_executor" />
     </list>
     <recent_temporary>
       <list>
-        <item itemvalue="Python.action_execute" />
+        <item itemvalue="Python.action_executor" />
         <item itemvalue="Python.keyboard_write_test" />
         <item itemvalue="Python.critical_exit_test" />
         <item itemvalue="Python.mouse_test" />
@@ -273,7 +273,7 @@
       <workItem from="1636090882090" duration="967000" />
       <workItem from="1636891087149" duration="1412000" />
       <workItem from="1636941910253" duration="1113000" />
-      <workItem from="1637030570122" duration="10330000" />
+      <workItem from="1637030570122" duration="18150000" />
     </task>
     <servers />
   </component>
@@ -320,7 +320,7 @@
     <SUITE FILE_PATH="coverage/Python_JEAutoControl$win32_listener.coverage" NAME="win32_listener Coverage Results" MODIFIED="1633687248883" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$/je_auto_control/windows/listener" />
     <SUITE FILE_PATH="coverage/Python_JEAutoControl$keyboard_write_test.coverage" NAME="keyboard_write_test Coverage Results" MODIFIED="1636891949186" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$/test/platform_independent/test/keyboard" />
     <SUITE FILE_PATH="coverage/Python_JEAutoControl$locate_image_center_test.coverage" NAME="locate_image_center_test Coverage Results" MODIFIED="1629389045129" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$/circle_ci_test" />
-    <SUITE FILE_PATH="coverage/Python_JEAutoControl$action_execute.coverage" NAME="action_execute Coverage Results" MODIFIED="1637075329164" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$/je_auto_control/utils/action_executer" />
+    <SUITE FILE_PATH="coverage/Python_JEAutoControl$action_execute.coverage" NAME="action_execute Coverage Results" MODIFIED="1637128387624" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$/je_auto_control/utils/action_executer" />
     <SUITE FILE_PATH="coverage/Python_JEAutoControl$json_action_test.coverage" NAME="json_action_test Coverage Results" MODIFIED="1634291087518" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$/test/platform_independent/test" />
     <SUITE FILE_PATH="coverage/Python_JEAutoControl$win32_record.coverage" NAME="win32_record Coverage Results" MODIFIED="1633892287671" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$/je_auto_control/windows/record" />
     <SUITE FILE_PATH="coverage/Python_JEAutoControl$record_test__1_.coverage" NAME="record_test (1) Coverage Results" MODIFIED="1634135762635" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$/circle_ci_test" />
