+++ je_auto_control/__init__.py
@@ -50,6 +50,6 @@ from je_auto_control.wrapper.auto_control_record import record
 from je_auto_control.wrapper.auto_control_record import stop_record
 
 # json
-from je_auto_control.utils.action_executer.action_execute import execute_action
-from je_auto_control.utils.action_file.json_loader import read_action_json
-from je_auto_control.utils.action_file.json_loader import write_action_json
+from je_auto_control.utils.action_executer.action_executor import execute_action
+from je_auto_control.utils.action_file.json_file import read_action_json
+from je_auto_control.utils.action_file.json_file import write_action_json
